feat: delete file an module settings adv
