+++ src/components/modals/SettingsAdv/SettingsAdv.module.css
@@ -117,22 +117,3 @@
     font-size: 16px;
     line-height: 150%;
 }
-
-.boxImg {
-    display: flex;
-    column-gap: 10px;
-}
-
-.img {
-    cursor: pointer;
-    width: 90px;
-    height: 90px;
-}
-
-.img:hover {
-    border: 2px solid #009ee4;
-}
-
-.img:active {
-    background-color: #eceaea;
-}
